+++ pet-adoption/pet-adoption-frontend/package.json
@@ -18,7 +18,8 @@
     "react-icons": "^5.3.0",
     "react-modal": "^3.16.1",
     "react-router-dom": "^6.27.0",
-    "react-select": "^5.8.1"
+    "react-select": "^5.8.1",
+    "react-toastify": "^10.0.6"
   },
   "devDependencies": {
     "@eslint/js": "^9.11.1",
